Practical task #02. Changed package's name mvp to mvvm.
