+++ app/build.gradle
@@ -56,6 +56,13 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:3.12.1'
     implementation 'com.jakewharton.retrofit:retrofit2-rxjava2-adapter:1.0.0'
 
+    // Dagger 2
+    implementation 'com.google.dagger:dagger:2.26'
+//    implementation 'com.google.dagger:dagger-android:2.17'
+//    implementation 'com.google.dagger:dagger-android-support:2.17'
+//    kapt 'com.google.dagger:dagger-android-processor:2.17'
+    kapt 'com.google.dagger:dagger-compiler:2.24'
+
     //Test
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test:runner:1.3.0'
