--- app/src/main/java/pt/svcdev/dictionary/mvp/view/base/View.kt
@@ -1,7 +1,0 @@
-package pt.svcdev.dictionary.mvp.view.base
-
-import pt.svcdev.dictionary.mvp.model.data.AppState
-
-interface View {
-    fun renderData(appState: AppState)
-}
