+++ app/src/main/java/pt/svcdev/dictionary/mvvm/model/datasource/DataSourceRemote.kt
@@ -1,7 +1,7 @@
-package pt.svcdev.dictionary.mvp.model.datasource
+package pt.svcdev.dictionary.mvvm.model.datasource
 
 import io.reactivex.Observable
-import pt.svcdev.dictionary.mvp.model.data.DataModel
+import pt.svcdev.dictionary.mvvm.model.data.DataModel
 
 class DataSourceRemote(private val remoteProvider: RetrofitImpl = RetrofitImpl()) :
     DataSource<List<DataModel>> {
