+++ app/src/main/java/pt/svcdev/dictionary/mvvm/view/base/View.kt
@@ -1,0 +1,7 @@
+package pt.svcdev.dictionary.mvvm.view.base
+
+import pt.svcdev.dictionary.mvvm.model.data.AppState
+
+interface View {
+    fun renderData(appState: AppState)
+}
