+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,11 @@
     <string name="button_search">Search</string>
     <string name="search_hint">New word</string>
     <string name="delete_description">X</string>
+    <string name="dialog_title_device_is_offline">No internet connection</string>
+    <string name="dialog_message_device_is_offline">
+        You can not use app without internet connection. Please, check internet connection.
+    </string>
+    <string name="dialog_tittle_sorry">Sorry</string>
+    <string name="dialog_button_cancel">Cancel</string>
+    <string name="dialog_title_stub">Undefined error</string>
 </resources>
